+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2022-04-20
+
+## Added
+
+- Cascade class
+- @cascading decorator
+
+### Changed
+
+- CircleCI workflows
+
 ## [0.1.1] - 2022-03-28
 
 ### Fixed